+++ coinlib/test/dlc/terms.dart
@@ -16,10 +16,10 @@ CETOutputs getOuts(List<String> coins) => CETOutputs(
 final exampleLocktime = MedianTimeLocktime(DateTime(2026));
 
 final exampleTerms = DLCTerms(
-  participants: [
+  participants: {
     getPubKey(0),
     getPubKey(1, false),
-  ],
+  },
   fundAmounts: {
     // Can be different from participants
     getPubKey(0): CoinUnit.coin.toSats("2"),
@@ -30,6 +30,19 @@ final exampleTerms = DLCTerms(
     getPubKey(4, false): getOuts(["6"]),
   },
   refundLocktime: exampleLocktime,
+  network: Network.mainnet,
+);
+
+DLCTerms getTerms({
+  Set<ECPublicKey>? participants,
+  Map<ECPublicKey, BigInt>? fundAmounts,
+  Map<ECPublicKey, CETOutputs>? outcomes,
+  Locktime? locktime,
+}) => DLCTerms(
+  participants: participants ?? exampleTerms.participants,
+  fundAmounts: fundAmounts ?? exampleTerms.fundAmounts,
+  outcomes: outcomes ?? exampleTerms.outcomes,
+  refundLocktime: locktime ?? exampleLocktime,
   network: Network.mainnet,
 );
 
@@ -64,7 +77,7 @@ void main() {
 
     test("immutable fields", () {
       expect(
-        () => exampleTerms.participants.first = examplePubkey,
+        () => exampleTerms.participants.add(examplePubkey),
         throwsA(anything),
       );
       expect(
@@ -78,25 +91,17 @@ void main() {
     });
 
     test("outcomes and funded amounts must match", () => expectInvalid(
-      () => DLCTerms(
-        participants: exampleTerms.participants,
+      () => getTerms(
         fundAmounts: { getPubKey(0): CoinUnit.coin.toSats("1") },
-        outcomes: exampleTerms.outcomes,
-        refundLocktime: exampleLocktime,
-        network: Network.mainnet,
       ),
     ),);
 
     test("funded amounts must be at least minOutput", () => expectInvalid(
-      () => DLCTerms(
-        participants: exampleTerms.participants,
+      () => getTerms(
         fundAmounts: {
           getPubKey(0): CoinUnit.coin.toSats("5.990001"),
           getPubKey(1): CoinUnit.coin.toSats("0.009999"),
         },
-        outcomes: exampleTerms.outcomes,
-        refundLocktime: exampleLocktime,
-        network: Network.mainnet,
       ),
     ),);
 
@@ -146,6 +151,53 @@ void main() {
       final bytes = exampleTerms.toBytes();
       bytes[0] = 0xff;
       expectInvalid(() => DLCTerms.fromBytes(bytes, Network.mainnet));
+    });
+
+    test("coerces keys into xonly", () {
+
+      final odd = ECPublicKey.fromHex(
+        "0379be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798",
+      );
+
+      expect(odd.yIsEven, false);
+
+      final terms = getTerms(
+        participants: { odd },
+        fundAmounts: { odd: CoinUnit.coin.toSats("1") },
+        outcomes: { odd: getOuts(["1"]) },
+      );
+
+      expect(terms.participants.first.yIsEven, true);
+      expect(terms.fundAmounts.keys.first.yIsEven, true);
+      expect(terms.outcomes.keys.first.yIsEven, true);
+
+    });
+
+    test("bad ordering of serialised keys", () {
+      final bytes = exampleTerms.toBytes();
+      // Swap first and second public keys
+      final first = bytes.sublist(3, 35);
+      bytes.setRange(3, 35, bytes, 35);
+      bytes.setAll(35, first);
+      expectInvalid(() => DLCTerms.fromBytes(bytes, Network.mainnet));
+    });
+
+    test("different musig for different data", () {
+
+      final musig1 = exampleTerms.musig;
+      final musig2 = getTerms(
+        locktime: MedianTimeLocktime(DateTime(2030)),
+      ).musig;
+
+      /// Same set of participant keys
+      for (final musig in [musig1, musig2]) {
+        expect(musig.pubKeys, hasLength(2));
+        expect(musig.pubKeys, containsAll(exampleTerms.participants));
+      }
+
+      // Different tweaked key
+      expect(musig1.aggregate, isNot(musig2.aggregate));
+
     });
 
   });